realizado alteração no codigo para responsividade; visualização da senha em login e afins

+++ src/views/motor/motorRegister.js
@@ -488,7 +488,7 @@ const MotorRegister = () => {
             ranhuras: parseInt(ranhuras),
             rotacao: parseInt(rotacao),
             ligacao,
-            potencia: parseInt(potencia),
+            potencia: parseFloat(potencia),
             comprimento: parseInt(comprimento),
             medidaInterna: parseInt(medidaInterna),
             empresa,
